navboard: remove wm board
There were bugs in either X, my wm or navboard tha caused problems when
navboard triggered a keybinding that was grabbed by my window and said
binding caused the wm to grab the keyboard. The result would be that the
key release wouldn't be seen the keyboard would be remained grabbed
instead of being released.

Instead of trying to trigger bindings with keypresses, just trigger the
action directly.