profile: Reduce HISTFILE when it gets too large
Want to keep a record of all commands that I could want to access in the
future. So want a large history file. However, I don't really care about
when they were used. So when the file gets too large, keep the last N
entries unmodified and run the rest of the file through `sort -u`. On
top of this we explicitly remove some commands that aren't interesting to
the history