phone: combine ringd and smsd
Instead of needing 2 daemons to handle processing calls and sms, the
phone project into requiring them. Instead both configs can be stored in
a ~/.config/phone and will be processed like normal. A slight side
effect is that they won't automatically see variables form ~/.profile
but if that ever becomes a problem, the file can be explicitly sourced.