Added xssh alias
Added an alias for X11 forwarding over ssh without needing xauth. This
works by taking advantage of ssh's -R option to forward the remote unix
domain socket corresponding to DISPLAY 2222 to the local display's
domain socket. Then in .profile, we default the DISPLAY variable to :2222

The main benefit of this is be able to copy/paste via xsel in vim easily